Offer a github commit which can be used as a dependency
Using a github commit as a dependency won't run the prepublish script, hence
won't compile the coffeescript, hence will make the package unusable.

This commit here can be used instead, since it contains all the compiled
code.  It ist NOT intended for the development of new releases, just an
intemediate solution until the next official release includes the fixes.